# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Internbot API",
    "version": "1.0.0",
    "description": "Backend workflow API for the Internbot internship platform. Source of truth: [WORKFLOW-API-SPEC.md](https://github.com/giatinhuynh/Internbot/blob/develop/docs/WORKFLOW-API-SPEC.md)."
  },
...
  "paths": {
...
    "/api/v1/opportunities": {
      "get": {
        "operationId": "listOpportunities",
        "summary": "List opportunities visible to the caller",
        "description": "Students are server-filtered to published opportunities in their selected semester. Coordinators can filter by semester, status, and type. See WORKFLOW-API-SPEC.md §7.3.",
        "tags": ["Opportunities"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "semesterId",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": ["draft", "pending_verification", "published", "rejected", "archived"]
              }
            }
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": ["pre_approved", "custom"]
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            }
          },
          {
            "in": "query",
            "name": "pageToken",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "description": "Default `-createdAt`. Prefix with `-` for descending.",
              "type": "string",
              "enum": ["createdAt", "-createdAt"]
            },
            "description": "Default `-createdAt`. Prefix with `-` for descending."
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of opportunities.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed query string or student attempted to widen semester filter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Firebase ID token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Student caller has no selected semester.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createOpportunity",
        "summary": "Create an opportunity",
        "description": "Coordinator-created opportunities start as draft. Student submissions become custom opportunities pending verification.",
        "tags": ["Opportunities"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOpportunityRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Opportunity created.",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed body or lifecycle fields supplied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Student has no selected semester or semester is not active.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Missing required fields or Career Hub URL not on allowlist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/opportunities/{id}": {
      "get": {
        "operationId": "getOpportunity",
        "summary": "Return an opportunity by id",
        "description": "Coordinators can read any opportunity. Students can read only published opportunities in their selected semester.",
        "tags": ["Opportunities"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "example": "opp_042",
              "description": "Platform opportunity id."
            },
            "required": true,
            "description": "Platform opportunity id."
          }
        ],
        "responses": {
          "200": {
            "description": "Opportunity found.",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityResponse"
                }
              }
            }
          },
          "403": {
            "description": "Student cannot see this opportunity.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No opportunity exists with the supplied id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "patchOpportunity",
        "summary": "Update opportunity metadata",
        "description": "Coordinator-only. Lifecycle fields are immutable through PATCH.",
        "tags": ["Opportunities"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "example": "opp_042",
              "description": "Platform opportunity id."
            },
            "required": true,
            "description": "Platform opportunity id."
          },
          {
            "in": "header",
            "name": "if-match",
            "schema": {
              "description": "Opt-in optimistic concurrency — the current ETag from a prior GET.",
              "type": "string"
            },
            "description": "Opt-in optimistic concurrency — the current ETag from a prior GET."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchOpportunityRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Opportunity updated.",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityResponse"
                }
              }
            }
          },
          "400": {
            "description": "Body contains immutable or unknown fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not a coordinator.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No opportunity exists with the supplied id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "412": {
            "description": "Stale `If-Match`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Empty body or domain validation failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/opportunities/{id}/attachments/{attachmentId}": {
      "get": {
        "operationId": "getOpportunityAttachment",
        "summary": "Return an opportunity attachment download resource",
        "description": "Returns attachment metadata with a fresh short-lived V4 signed Cloud Storage URL. Students can read only attachments on visible opportunities.",
        "tags": ["Opportunities"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "example": "opp_042",
              "description": "Platform opportunity id."
            },
            "required": true,
            "description": "Platform opportunity id."
          },
          {
            "in": "path",
            "name": "attachmentId",
            "schema": {
              "type": "string",
              "example": "att_001",
              "description": "Attachment id."
            },
            "required": true,
            "description": "Attachment id."
          }
        ],
        "responses": {
          "200": {
            "description": "Attachment found with a fresh signed download URL.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityAttachmentDownloadResponse"
                }
              }
            }
          },
          "403": {
            "description": "Student cannot see the parent opportunity.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No opportunity or attachment exists with the supplied id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/opportunities/{id}/transitions": {
      "post": {
        "operationId": "transitionOpportunity",
        "summary": "Advance an opportunity's coordinator-controlled status",
        "description": "Coordinator-only. Allowed: draft → published, draft → archived, published → archived. Pending-verification opportunities use /verifications.",
        "tags": ["Opportunities"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "example": "opp_042",
              "description": "Platform opportunity id."
            },
            "required": true,
            "description": "Platform opportunity id."
          },
          {
            "in": "header",
            "name": "if-match",
            "schema": {
              "description": "Opt-in optimistic concurrency — the current ETag from a prior GET.",
              "type": "string"
            },
            "description": "Opt-in optimistic concurrency — the current ETag from a prior GET."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionOpportunityRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transition applied; activity record written.",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityResponse"
                }
              }
            }
          },
          "409": {
            "description": "Disallowed state transition.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "412": {
            "description": "Stale `If-Match`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/opportunities/{id}/verifications": {
      "post": {
        "operationId": "verifyOpportunity",
        "summary": "Verify or reject a student-submitted opportunity",
        "description": "Coordinator-only. Applies only to pending_verification opportunities and creates a notification for the submitter.",
        "tags": ["Opportunities"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "example": "opp_042",
              "description": "Platform opportunity id."
            },
            "required": true,
            "description": "Platform opportunity id."
          },
          {
            "in": "header",
            "name": "if-match",
            "schema": {
              "description": "Opt-in optimistic concurrency — the current ETag from a prior GET.",
              "type": "string"
            },
            "description": "Opt-in optimistic concurrency — the current ETag from a prior GET."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOpportunityRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Verification applied.",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                }
              },
              "ETag": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityResponse"
                }
              }
            }
          },
          "409": {
            "description": "Opportunity is not pending verification.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "412": {
            "description": "Stale `If-Match`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Rejected verification missing comment.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateOpportunityRequest": {
        "type": "object",
        "properties": {
          "semesterId": {
            "type": "string",
            "minLength": 1
          },
          "type": {
            "type": "string",
            "enum": ["pre_approved", "custom"]
          },
          "employerName": {
            "type": "string",
            "minLength": 1,
            "example": "Example Pty Ltd"
          },
          "jobTitle": {
            "type": "string",
            "minLength": 1,
            "example": "Software Intern"
          },
          "descriptionText": {
            "type": "string",
            "minLength": 1
          },
          "workMode": {
            "type": "string",
            "enum": ["onsite", "hybrid", "remote"]
          },
          "location": {
            "type": "string",
            "minLength": 1
          },
          "sourceUrl": {
            "example": "https://careerhub.rmit.edu.au/jobs/12345",
            "type": "string",
            "format": "uri"
          }
        },
        "required": ["employerName", "jobTitle", "descriptionText"],
        "additionalProperties": false,
        "description": "Body for POST /api/v1/opportunities. Coordinator and student callers have different server-side defaults. See WORKFLOW-API-SPEC.md §7.3."
      },
      "PatchOpportunityRequest": {
        "type": "object",
        "properties": {
          "employerName": {
            "type": "string",
            "minLength": 1
          },
          "jobTitle": {
            "type": "string",
            "minLength": 1
          },
          "descriptionText": {
            "type": "string",
            "minLength": 1
          },
          "workMode": {
            "anyOf": [
              {
                "type": "string",
                "enum": ["onsite", "hybrid", "remote"]
              },
              {
                "type": "null"
              }
            ]
          },
          "location": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ]
          },
          "sourceUrl": {
            "anyOf": [
              {
                "type": "string",
                "format": "uri"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "description": "Body for PATCH /api/v1/opportunities/:id. Lifecycle fields are rejected with 400. See WORKFLOW-API-SPEC.md §7.3."
      },
      "TransitionOpportunityRequest": {
        "type": "object",
        "properties": {
          "to": {
            "type": "string",
            "enum": ["published", "archived"]
          },
          "comment": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": ["to"],
        "additionalProperties": false,
        "description": "Body for POST /api/v1/opportunities/:id/transitions. See WORKFLOW-API-SPEC.md §7.3."
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "https://httpstatuses.io/403"
          },
          "title": {
            "type": "string",
            "example": "Forbidden"
          },
          "status": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "detail": {
            "type": "string"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": ["type", "title", "status", "detail", "error"],
        "additionalProperties": false,
        "description": "RFC 9457 Problem Details wrapped with the domain error envelope — see docs/ERROR-HANDLING.md."
      },
...
      "OpportunityListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OpportunityResponse"
            }
          },
          "nextPageToken": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": ["items", "nextPageToken"],
        "additionalProperties": false,
        "description": "Paginated list of opportunities."
      },
...
      "OpportunityAttachmentDownloadResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "att_001"
          },
          "fileName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "example": "position-description.pdf"
          },
          "contentType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "example": "application/pdf"
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
            "example": "2026-04-04T09:05:00Z"
          },
          "downloadUrl": {
            "type": "string",
            "format": "uri"
          },
          "downloadUrlExpiresAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          }
        },
        "required": [
          "id",
          "fileName",
          "contentType",
          "uploadedAt",
          "downloadUrl",
          "downloadUrlExpiresAt"
        ],
        "additionalProperties": false,
        "description": "Opportunity attachment metadata with a fresh short-lived Cloud Storage signed URL."
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase ID token",
        "description": "Firebase ID token obtained client-side. Platform identity is hydrated at the api edge from `userIdentities/{provider}__{uid}` → `users/{id}` on every request; on a brand-new student-shape email the hydrator JIT-creates the platform record in the same transaction. See docs/WORKFLOW-API-SPEC.md §7.0."
      }
    }
  }
}